Organizing Favourites and separating the logic and UI into pieces inside Controller
