+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Favorites/FavoritesViewController.swift
@@ -12,54 +12,27 @@ import Shared
 import UIKit
 import os.log
 
-private class FavoritesHeaderView: UICollectionReusableView {
-  let label = UILabel().then {
-    $0.text = Strings.recentSearchFavorites
-    $0.font = .systemFont(ofSize: 18, weight: .semibold)
-  }
-  override init(frame: CGRect) {
-    super.init(frame: frame)
-    addSubview(label)
-    label.snp.makeConstraints {
-      $0.leading.equalToSuperview().inset(12)
-      $0.trailing.lessThanOrEqualToSuperview().inset(12)
-      $0.centerY.equalToSuperview()
-    }
-  }
-  @available(*, unavailable)
-  required init(coder: NSCoder) {
-    fatalError()
-  }
-}
-
 class FavoritesViewController: UIViewController {
 
-  /// CoreData's diffable method relies on managed object ID only.
-  /// This does not work in our case since the object may be the same but with changed title or order.
-  /// This struct stores managed object ID as well as properties that we observer whether they have changed.
-  /// Note: `order` property does not have to be stored, favorite's frc handles order updates, it returns items in correct order.
-  private struct FavoriteDiffable: Hashable {
-    let objectID: NSManagedObjectID
-    let title: String?
-    let url: String?
+  // UI Properties
+  let collectionView: UICollectionView
+  private let backgroundView = UIVisualEffectView(
+    effect: UIBlurEffect(style: .systemUltraThinMaterial)
+  )
 
-    func hash(into hasher: inout Hasher) {
-      hasher.combine(objectID)
-    }
+  private let layout = UICollectionViewFlowLayout().then {
+    $0.sectionInset = UIEdgeInsets(top: 12, left: 0, bottom: 22, right: 0)
+    $0.minimumInteritemSpacing = 0
+    $0.minimumLineSpacing = 8
   }
 
-  /// Favorites VC has two fetch result controllers to pull from.
-  /// This enum stores both models.
-  private enum DataWrapper: Hashable {
-    case favorite(FavoriteDiffable)
+  private var favoriteGridSize: CGSize = .zero
 
-    // Recent searches are static, we do not need any wrapper class for them.
-    case recentSearch(NSManagedObjectID)
-  }
-
-  private typealias DataSource = UICollectionViewDiffableDataSource<Section, DataWrapper>
-  private typealias Snapshot = NSDiffableDataSourceSnapshot<Section, DataWrapper>
-
+  // Data Source
+  private typealias DataSource = UICollectionViewDiffableDataSource<
+    FavoritesSection, FavoritesDataWrapper
+  >
+  private typealias Snapshot = NSDiffableDataSourceSnapshot<FavoritesSection, FavoritesDataWrapper>
   private lazy var dataSource =
     DataSource(
       collectionView: self.collectionView,
@@ -69,71 +42,54 @@ class FavoritesViewController: UIViewController {
       }
     )
 
-  var action: (Favorite, BookmarksAction) -> Void
+  // Actions
+  var bookmarkAction: (Favorite, BookmarksAction) -> Void
   var recentSearchAction: (RecentSearch?, Bool) -> Void
 
-  private enum Section: Int, CaseIterable {
-    case pasteboard = 0
-    case favorites = 1
-    case recentSearches = 2
-  }
-
-  private let favoritesFRC = Favorite.frc()
+  // Fetched Result
+  let favoritesFRC = Favorite.frc()
   private let recentSearchesFRC = RecentSearch.frc().then {
     $0.fetchRequest.fetchLimit = 5
   }
 
-  private let layout = UICollectionViewFlowLayout().then {
-    $0.sectionInset = UIEdgeInsets(top: 12, left: 0, bottom: 22, right: 0)
-    $0.minimumInteritemSpacing = 0
-    $0.minimumLineSpacing = 8
-  }
-
-  private var tabType: TabType
-  private var favoriteGridSize: CGSize = .zero
-  private let collectionView: UICollectionView
-  private let backgroundView = UIVisualEffectView(
-    effect: UIBlurEffect(style: .systemUltraThinMaterial)
-  )
-  private var hasPasteboardURL = false
-  private var privateBrowsingManager: PrivateBrowsingManager
+  // Private Browsing
+  var privateBrowsingManager: PrivateBrowsingManager
   private var privateModeCancellable: AnyCancellable?
 
   init(
-    tabType: TabType,
     privateBrowsingManager: PrivateBrowsingManager,
-    action: @escaping (Favorite, BookmarksAction) -> Void,
+    bookmarkAction: @escaping (Favorite, BookmarksAction) -> Void,
     recentSearchAction: @escaping (RecentSearch?, Bool) -> Void
   ) {
-    self.tabType = tabType
-    self.action = action
+    self.bookmarkAction = bookmarkAction
     self.recentSearchAction = recentSearchAction
     self.privateBrowsingManager = privateBrowsingManager
     collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
 
     super.init(nibName: nil, bundle: nil)
 
-    collectionView.register(
-      RecentSearchClipboardHeaderView.self,
-      forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
-      withReuseIdentifier: "pasteboard_header"
-    )
-    collectionView.register(FavoriteCell.self)
-    collectionView.register(RecentSearchCell.self)
-    collectionView.register(
-      FavoritesHeaderView.self,
-      forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
-      withReuseIdentifier: "fav_header"
-    )
-    collectionView.register(
-      RecentSearchHeaderView.self,
-      forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
-      withReuseIdentifier: "recent_searches_header"
-    )
+    collectionView.do {
+      $0.register(
+        FavoriteRecentSearchClipboardHeaderView.self,
+        forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
+        withReuseIdentifier: "pasteboard_header"
+      )
+      $0.register(FavoritesCell.self)
+      $0.register(FavoritesRecentSearchCell.self)
+      $0.register(
+        FavoritesHeaderView.self,
+        forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
+        withReuseIdentifier: "fav_header"
+      )
+      $0.register(
+        FavoritesRecentSearchHeaderView.self,
+        forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
+        withReuseIdentifier: "recent_searches_header"
+      )
+    }
 
     favoritesFRC.delegate = self
     recentSearchesFRC.delegate = self
-    hasPasteboardURL = UIPasteboard.general.hasStrings || UIPasteboard.general.hasURLs
 
     KeyboardHelper.defaultHelper.addDelegate(self)
 
@@ -158,28 +114,6 @@ class FavoritesViewController: UIViewController {
   override func viewDidLoad() {
     super.viewDidLoad()
 
-    view.addSubview(backgroundView)
-    view.addSubview(collectionView)
-
-    backgroundView.snp.makeConstraints {
-      $0.edges.equalToSuperview()
-    }
-    collectionView.snp.makeConstraints {
-      $0.edges.equalToSuperview()
-    }
-
-    collectionView.alwaysBounceVertical = true
-    collectionView.contentInset = UIEdgeInsets(top: 24, left: 0, bottom: 0, right: 0)
-    collectionView.backgroundColor = .clear
-    collectionView.dataSource = dataSource
-    collectionView.delegate = self
-    collectionView.dragDelegate = self
-    collectionView.dropDelegate = self
-
-    // Drag should be enabled to rearrange favourite
-    collectionView.dragInteractionEnabled = true
-    collectionView.keyboardDismissMode = .interactive
-
     dataSource.supplementaryViewProvider = { [weak self] collectionView, kind, indexPath in
       self?.supplementaryViewProvider(
         collectionView: collectionView,
@@ -188,8 +122,10 @@ class FavoritesViewController: UIViewController {
       )
     }
 
+    doLayout()
+    setTheme()
+
     updateUIWithSnapshot()
-    updateColors()
   }
 
   override func viewDidLayoutSubviews() {
@@ -216,6 +152,36 @@ class FavoritesViewController: UIViewController {
     calculateAppropriateGrid()
   }
 
+  private func doLayout() {
+    view.addSubview(backgroundView)
+    view.addSubview(collectionView)
+
+    backgroundView.snp.makeConstraints {
+      $0.edges.equalToSuperview()
+    }
+    collectionView.snp.makeConstraints {
+      $0.edges.equalToSuperview()
+    }
+  }
+
+  private func setTheme() {
+    collectionView.do {
+      $0.alwaysBounceVertical = true
+      $0.contentInset = UIEdgeInsets(top: 24, left: 0, bottom: 0, right: 0)
+      $0.backgroundColor = .clear
+      $0.dataSource = dataSource
+      $0.delegate = self
+      $0.dragDelegate = self
+      $0.dropDelegate = self
+
+      // Drag should be enabled to rearrange favourite
+      $0.dragInteractionEnabled = true
+      $0.keyboardDismissMode = .interactive
+    }
+
+    updateColors()
+  }
+
   private func updateColors() {
     let browserColors = privateBrowsingManager.browserColors
     // Have to apply a custom alpha here because UIVisualEffectView blurs come with their own tint
@@ -232,17 +198,17 @@ class FavoritesViewController: UIViewController {
     let minimumNumberOfColumns = 4
     let minWidth = floor(width / CGFloat(minimumNumberOfColumns))
     // Default width should be 82, but may get smaller or bigger
-    var itemSize = CGSize(width: 82, height: FavoriteCell.height(forWidth: 82))
+    var itemSize = CGSize(width: 82, height: FavoritesCell.height(forWidth: 82))
     if minWidth < 82 {
       itemSize = CGSize(
         width: floor(width / 4.0),
-        height: FavoriteCell.height(forWidth: floor(width / 4.0))
+        height: FavoritesCell.height(forWidth: floor(width / 4.0))
       )
     } else if traitCollection.horizontalSizeClass == .regular {
       // On iPad's or Max/Plus phones allow the icons to get bigger to an
       // extent
       if width / CGFloat(minimumNumberOfColumns) > 100.0 {
-        itemSize = CGSize(width: 100, height: FavoriteCell.height(forWidth: 100))
+        itemSize = CGSize(width: 100, height: FavoritesCell.height(forWidth: 100))
       }
     }
 
@@ -261,63 +227,8 @@ class FavoritesViewController: UIViewController {
   }
 }
 
-// MARK: - KeyboardHelperDelegate
-extension FavoritesViewController: KeyboardHelperDelegate {
-  func updateKeyboardInset(_ state: KeyboardState, animated: Bool = true) {
-    if collectionView.bounds.size == .zero { return }
-    let keyboardHeight =
-      state.intersectionHeightForView(self.view) - view.safeAreaInsets.bottom
-      + additionalSafeAreaInsets.bottom
-    UIViewPropertyAnimator(
-      duration: animated ? state.animationDuration : 0.0,
-      curve: state.animationCurve
-    ) {
-      self.collectionView.contentInset = self.collectionView.contentInset.with {
-        $0.bottom = keyboardHeight
-      }
-      self.collectionView.scrollIndicatorInsets = self.collectionView.verticalScrollIndicatorInsets
-        .with {
-          $0.bottom = keyboardHeight
-        }
-    }.startAnimation()
-  }
-
-  func keyboardHelper(
-    _ keyboardHelper: KeyboardHelper,
-    keyboardWillShowWithState state: KeyboardState
-  ) {
-    updateKeyboardInset(state)
-  }
-
-  func keyboardHelper(
-    _ keyboardHelper: KeyboardHelper,
-    keyboardWillHideWithState state: KeyboardState
-  ) {
-    updateKeyboardInset(state)
-  }
-
-  private var availableSections: [Section] {
-    var sections = [Section]()
-    if hasPasteboardURL {
-      sections.append(.pasteboard)
-    }
-
-    if let favoritesObjects = favoritesFRC.fetchedObjects, !favoritesObjects.isEmpty {
-      sections.append(.favorites)
-    }
-
-    if !tabType.isPrivate && Preferences.Search.shouldShowRecentSearches.value,
-      RecentSearch.totalCount() > 0
-    {
-      sections.append(.recentSearches)
-    } else if !tabType.isPrivate && Preferences.Search.shouldShowRecentSearchesOptIn.value {
-      sections.append(.recentSearches)
-    }
-    return sections
-  }
-}
-
 // MARK: - UICollectionViewDataSource & UICollectionViewDelegateFlowLayout
+
 extension FavoritesViewController: UICollectionViewDelegateFlowLayout {
 
   func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
@@ -333,7 +244,7 @@ extension FavoritesViewController: UICollectionViewDelegateFlowLayout {
       guard let bookmark = favoritesFRC.fetchedObjects?[safe: indexPath.item] else {
         return
       }
-      action(bookmark, .opened())
+      bookmarkAction(bookmark, .opened())
     case .recentSearches:
       guard let searchItem = recentSearchesFRC.fetchedObjects?[safe: indexPath.item] else {
         return
@@ -397,204 +308,6 @@ extension FavoritesViewController: UICollectionViewDelegateFlowLayout {
         - (collectionView.contentInset.left + collectionView.contentInset.right)
       return CGSize(width: width, height: 28.0)
     }
-  }
-
-  func collectionView(
-    _ collectionView: UICollectionView,
-    contextMenuConfigurationForItemAt indexPath: IndexPath,
-    point: CGPoint
-  ) -> UIContextMenuConfiguration? {
-    guard let section = availableSections[safe: indexPath.section] else {
-      assertionFailure("Invalid Section")
-      return nil
-    }
-
-    switch section {
-    case .pasteboard:
-      break
-    case .favorites:
-      guard let bookmark = favoritesFRC.fetchedObjects?[indexPath.item] else { return nil }
-      return UIContextMenuConfiguration(identifier: indexPath as NSCopying, previewProvider: nil) {
-        _ -> UIMenu? in
-        let openInNewTab = UIAction(
-          title: Strings.openNewTabButtonTitle,
-          handler: UIAction.deferredActionHandler { _ in
-            self.action(bookmark, .opened(inNewTab: true, switchingToPrivateMode: false))
-          }
-        )
-        let edit = UIAction(
-          title: Strings.editFavorite,
-          handler: UIAction.deferredActionHandler { _ in
-            self.action(bookmark, .edited)
-          }
-        )
-        let delete = UIAction(
-          title: Strings.removeFavorite,
-          attributes: .destructive,
-          handler: UIAction.deferredActionHandler { _ in
-            bookmark.delete()
-          }
-        )
-
-        var urlChildren: [UIAction] = [openInNewTab]
-        if !self.privateBrowsingManager.isPrivateBrowsing {
-          let openInNewPrivateTab = UIAction(
-            title: Strings.openNewPrivateTabButtonTitle,
-            handler: UIAction.deferredActionHandler { _ in
-              self.action(bookmark, .opened(inNewTab: true, switchingToPrivateMode: true))
-            }
-          )
-          urlChildren.append(openInNewPrivateTab)
-        }
-
-        let urlMenu = UIMenu(title: "", options: .displayInline, children: urlChildren)
-        let favMenu = UIMenu(title: "", options: .displayInline, children: [edit, delete])
-        return UIMenu(
-          title: bookmark.title ?? bookmark.url ?? "",
-          identifier: nil,
-          children: [urlMenu, favMenu]
-        )
-      }
-    case .recentSearches:
-      break
-    }
-    return nil
-  }
-
-  func collectionView(
-    _ collectionView: UICollectionView,
-    previewForHighlightingContextMenuWithConfiguration configuration: UIContextMenuConfiguration
-  ) -> UITargetedPreview? {
-    guard let indexPath = configuration.identifier as? IndexPath,
-      let cell = collectionView.cellForItem(at: indexPath) as? FavoriteCell
-    else {
-      return nil
-    }
-    return UITargetedPreview(view: cell.imageView)
-  }
-
-  func collectionView(
-    _ collectionView: UICollectionView,
-    previewForDismissingContextMenuWithConfiguration configuration: UIContextMenuConfiguration
-  ) -> UITargetedPreview? {
-    guard let indexPath = configuration.identifier as? IndexPath,
-      let cell = collectionView.cellForItem(at: indexPath) as? FavoriteCell
-    else {
-      return nil
-    }
-    return UITargetedPreview(view: cell.imageView)
-  }
-}
-
-// MARK: - UICollectionViewDragDelegate & UICollectionViewDropDelegate
-extension FavoritesViewController: UICollectionViewDragDelegate, UICollectionViewDropDelegate {
-  func collectionView(
-    _ collectionView: UICollectionView,
-    itemsForBeginning session: UIDragSession,
-    at indexPath: IndexPath
-  ) -> [UIDragItem] {
-    guard let section = availableSections[safe: indexPath.section] else {
-      assertionFailure("Invalid Section")
-      return []
-    }
-
-    switch section {
-    case .pasteboard:
-      break
-    case .favorites:
-      // Fetch results controller indexpath is independent from our collection view.
-      // All results of it are stored in first section.
-      let adjustedIndexPath = IndexPath(row: indexPath.row, section: 0)
-      let bookmark = favoritesFRC.object(at: adjustedIndexPath)
-      let itemProvider = NSItemProvider(object: "\(indexPath)" as NSString)
-      let dragItem = UIDragItem(itemProvider: itemProvider)
-      dragItem.previewProvider = { () -> UIDragPreview? in
-        guard let cell = collectionView.cellForItem(at: indexPath) as? FavoriteCell else {
-          return nil
-        }
-        return UIDragPreview(view: cell.imageView)
-      }
-      dragItem.localObject = bookmark
-      return [dragItem]
-    case .recentSearches:
-      break
-    }
-    return []
-  }
-
-  func collectionView(
-    _ collectionView: UICollectionView,
-    performDropWith coordinator: UICollectionViewDropCoordinator
-  ) {
-    guard let sourceIndexPath = coordinator.items.first?.sourceIndexPath else { return }
-    let destinationIndexPath: IndexPath
-    if let indexPath = coordinator.destinationIndexPath {
-      destinationIndexPath = indexPath
-    } else {
-      let section = max(collectionView.numberOfSections - 1, 0)
-      let row = collectionView.numberOfItems(inSection: section)
-      destinationIndexPath = IndexPath(row: max(row - 1, 0), section: section)
-    }
-
-    if sourceIndexPath.section != destinationIndexPath.section {
-      return
-    }
-
-    switch coordinator.proposal.operation {
-    case .move:
-      guard let item = coordinator.items.first else { return }
-      _ = coordinator.drop(item.dragItem, toItemAt: destinationIndexPath)
-      Favorite.reorder(
-        sourceIndexPath: sourceIndexPath,
-        destinationIndexPath: destinationIndexPath,
-        isInteractiveDragReorder: true
-      )
-    case .copy:
-      break
-    default: return
-    }
-  }
-
-  func collectionView(
-    _ collectionView: UICollectionView,
-    dropSessionDidUpdate session: UIDropSession,
-    withDestinationIndexPath destinationIndexPath: IndexPath?
-  ) -> UICollectionViewDropProposal {
-    if favoritesFRC.fetchedObjects?.count == 1 {
-      return .init(operation: .cancel)
-    }
-    return .init(operation: .move, intent: .insertAtDestinationIndexPath)
-  }
-
-  func collectionView(
-    _ collectionView: UICollectionView,
-    dragPreviewParametersForItemAt indexPath: IndexPath
-  ) -> UIDragPreviewParameters? {
-    let params = UIDragPreviewParameters()
-    params.backgroundColor = .clear
-    if let cell = collectionView.cellForItem(at: indexPath) as? FavoriteCell {
-      params.visiblePath = UIBezierPath(roundedRect: cell.imageView.frame, cornerRadius: 8)
-    }
-    return params
-  }
-
-  func collectionView(
-    _ collectionView: UICollectionView,
-    dropPreviewParametersForItemAt indexPath: IndexPath
-  ) -> UIDragPreviewParameters? {
-    let params = UIDragPreviewParameters()
-    params.backgroundColor = .clear
-    if let cell = collectionView.cellForItem(at: indexPath) as? FavoriteCell {
-      params.visiblePath = UIBezierPath(roundedRect: cell.imageView.frame, cornerRadius: 8)
-    }
-    return params
-  }
-
-  func collectionView(
-    _ collectionView: UICollectionView,
-    dragSessionIsRestrictedToDraggingApplication session: UIDragSession
-  ) -> Bool {
-    return true
   }
 }
 
@@ -662,6 +375,7 @@ extension FavoritesViewController: PreferencesObserver {
 }
 
 // MARK: - Diffable data source + NSFetchedResultsControllerDelegate
+
 extension FavoritesViewController: NSFetchedResultsControllerDelegate {
   private var favoritesSectionExists: Bool {
     availableSections.contains(.favorites)
@@ -683,8 +397,10 @@ extension FavoritesViewController: NSFetchedResultsControllerDelegate {
     var snapshot = Snapshot()
     snapshot.appendSections(availableSections)
 
-    let fetchedFavorites: [DataWrapper]? = favoritesFRC.fetchedObjects?.compactMap {
-      return .favorite(FavoriteDiffable(objectID: $0.objectID, title: $0.displayTitle, url: $0.url))
+    let fetchedFavorites: [FavoritesDataWrapper]? = favoritesFRC.fetchedObjects?.compactMap {
+      return .favorite(
+        FavoritesDiffable(objectID: $0.objectID, title: $0.displayTitle, url: $0.url)
+      )
     }
 
     if favoritesSectionExists, let favorites = fetchedFavorites {
@@ -718,7 +434,7 @@ extension FavoritesViewController: NSFetchedResultsControllerDelegate {
     guard let ids = snapshot.itemIdentifiers as? [NSManagedObjectID] else { return }
 
     if controller === favoritesFRC {
-      var items = [DataWrapper]()
+      var items = [FavoritesDataWrapper]()
 
       if favoritesSectionExists {
         ids.forEach {
@@ -728,7 +444,7 @@ extension FavoritesViewController: NSFetchedResultsControllerDelegate {
           if let existingItem = controller.managedObjectContext.object(with: $0) as? Favorite {
             items.append(
               .favorite(
-                FavoriteDiffable(
+                FavoritesDiffable(
                   objectID: $0,
                   title: existingItem.title,
                   url: existingItem.url
@@ -760,7 +476,7 @@ extension FavoritesViewController: NSFetchedResultsControllerDelegate {
       }
 
       if recentSearchesSectionExists {
-        var items = [DataWrapper]()
+        var items = [FavoritesDataWrapper]()
 
         ids.forEach {
           if RecentSearch.get(with: $0) != nil {
@@ -785,14 +501,14 @@ extension FavoritesViewController: NSFetchedResultsControllerDelegate {
   private func cellProvider(
     collectionView: UICollectionView,
     indexPath: IndexPath,
-    wrapper: DataWrapper
+    wrapper: FavoritesDataWrapper
   ) -> UICollectionViewCell? {
 
     switch wrapper {
     case .favorite(let favoriteWrapper):
       guard let favorite = Favorite.get(with: favoriteWrapper.objectID) else { return nil }
 
-      let cell = collectionView.dequeueReusableCell(for: indexPath) as FavoriteCell
+      let cell = collectionView.dequeueReusableCell(for: indexPath) as FavoritesCell
 
       cell.textLabel.text = favorite.displayTitle ?? favorite.url
       if let url = favorite.url?.asURL {
@@ -809,7 +525,7 @@ extension FavoritesViewController: NSFetchedResultsControllerDelegate {
     case .recentSearch(let objectID):
       guard let recentSearch = RecentSearch.get(with: objectID) else { return nil }
 
-      let cell = collectionView.dequeueReusableCell(for: indexPath) as RecentSearchCell
+      let cell = collectionView.dequeueReusableCell(for: indexPath) as FavoritesRecentSearchCell
 
       guard let searchType = RecentSearchType(rawValue: recentSearch.searchType) else {
         cell.setTitle(recentSearch.text)
@@ -890,7 +606,7 @@ extension FavoritesViewController: NSFetchedResultsControllerDelegate {
           ofKind: kind,
           withReuseIdentifier: "pasteboard_header",
           for: indexPath
-        ) as? RecentSearchClipboardHeaderView {
+        ) as? FavoriteRecentSearchClipboardHeaderView {
           header.button.removeTarget(self, action: nil, for: .touchUpInside)
           header.button.addTarget(self, action: #selector(onPasteboardAction), for: .touchUpInside)
           return header
@@ -908,7 +624,7 @@ extension FavoritesViewController: NSFetchedResultsControllerDelegate {
           ofKind: kind,
           withReuseIdentifier: "recent_searches_header",
           for: indexPath
-        ) as? RecentSearchHeaderView {
+        ) as? FavoritesRecentSearchHeaderView {
           header.resetLayout(showRecentSearches: Preferences.Search.shouldShowRecentSearches.value)
 
           header.showButton.addTarget(
