+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Search/RecentSearchQRCodeScannerController.swift
